+++ CHANGES.rst
@@ -125,6 +125,7 @@ Accessibility
 ^^^^^^^^^^^^^
 
 - Improve registration form date picker accessibility (:pr:`6371`, thanks :user:`foxbunny`)
+- Improve registration form single choice input accessibility (:pr:`6310`, thanks :user:`foxbunny`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^
@@ -136,6 +137,8 @@ Internal Changes
   (:pr:`6409`, thanks :user:`omegak`)
 - Custom poster/badge designer placeholder returning images need to return a ``BytesIO``
   instead of a Pillow ``Image`` object (:pr:`6441`)
+- Add ``<ind-date-picker>`` custom element (:pr:`6371`, thanks :user:`foxbunny`)
+- Add ``<ind-combobox>`` custom element (:pr:`6310`, thanks :user:`foxbunny`)
 
 
 Version 3.3.2